feat: add mailbox types
